pyproject.toml
[tool.poetry]
name = "tour"
version = "0.1.0"
description = ""
authors = ["marcocapozzoli <marcocapozzoli90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
psycopg2 = "^2.9.6"
pytest = "^7.3.1"
djangorestframework = "^3.14.0"
dj-database-url = "^2.0.0"
validate-docbr = "^1.10.0"
pydantic = {extras = ["email"], version = "^1.10.8"}
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.2.2"
isort = "^5.12.0"
black = "^23.3.0"
flake8 = "^6.0.0"
coverage = "^7.2.6"
pytest-cov = "^4.1.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
